refactor(evmcrispr): move aragonos functions to its own module
